pyproject.toml
[project]
name = "rag-orchestrator-agent"
version = "0.1.0"
description = "RAG 멀티에이전트 오케스트레이션 + MCP 기반 문서 DB"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "python-multipart>=0.0.9",
  "chromadb>=0.5.0",
  "mcp>=1.2.0",
  "pydantic>=2.0",
  "httpx>=0.27.0",
  "openai>=1.0.0",
  "sqlalchemy>=2.0.0",
  "python-dotenv>=1.0.0",
  "python-docx>=1.1.0",
  "openpyxl>=3.1.0",
  "python-pptx>=1.0.0",
  "langgraph>=0.2.0",
]

[project.optional-dependencies]
dev = ["ruff>=0.8.0"]

[tool.setuptools.packages.find]
where = ["."]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"
